# backend/storage/app/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'JourneyPlanner API Documentation'
  description: 'JourneyPlanner API - for internal use only'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting /token.'
security:
  -
    default: []
...
paths:
...
  '/api/journey/{journey_id}/regenerate-invite':
    post:
      summary: 'Regenerate invite'
      operationId: regenerateInvite
      description: 'Regenerate the invite code for the specified journey.'
      parameters: []
      responses:
        200:
          description: Success.
          content:
            application/json:
              schema:
                type: object
                example:
                  invite: d5b960bc-e0d4-463d-ac08-343803c05399
                properties:
                  invite:
                    type: string
                    example: d5b960bc-e0d4-463d-ac08-343803c05399
                    description: 'The new invite code.'
                    enum: []
                required:
                  - invite
        403:
          description: Unauthorized.
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'This action is unauthorized.'
                properties:
                  message:
                    type: string
                    example: 'This action is unauthorized.'
        404:
          description: 'Journey not found.'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'No query results for model [App\Models\Journey] 9e807de1-4b8e-4cbd-9dc5-cf2addb6a89b'
                properties:
                  message:
                    type: string
                    example: 'No query results for model [App\Models\Journey] 9e807de1-4b8e-4cbd-9dc5-cf2addb6a89b'
      tags:
        - Journey
    parameters:
      -
        in: path
        name: journey_id
        description: 'The ID of the journey.'
        example: 0023df66-7dfa-4009-afb0-57601eaa717c
        required: true
        schema:
          type: string
      -
        in: path
        name: id
        description: 'The UUID of the journey.'
        example: 9c23279f-f2a7-4666-bc00-56dfe0c47342
        required: true
        schema:
          type: uuid
...
  '/api/journey/{journey_id}/weather':
    get:
      summary: 'Get weather'
      operationId: getWeather
      description: 'Get the weather data for the specified journey.'
      parameters: []
      responses:
        200:
          description: Success.
          content:
            application/json:
              schema:
                type: object
                example:
                  current:
                    temperature: 12.6
                    weather_code: 2
                  forecast:
                    -
                      date: '2025-03-23'
                      temperature_max: 13.9
                      temperature_min: 7.7
                      weather_code: 80
                    -
                      date: '2025-03-24'
                      temperature_max: 15.9
                      temperature_min: 7.8
                      weather_code: 80
                    -
                      date: '2025-03-25'
                      temperature_max: 15.7
                      temperature_min: 7.6
                      weather_code: 3
                    -
                      date: '2025-03-26'
                      temperature_max: 15.9
                      temperature_min: 8.7
                      weather_code: 3
                properties:
                  current:
                    type: object
                    properties:
                      temperature:
                        type: number
                        example: 12.6
                        description: 'The current temperature.'
                        enum: []
                      weather_code:
                        type: integer
                        example: 2
                        description: 'The current weather code.'
                        enum: []
                    required:
                      - temperature
                      - weather_code
                    description: 'The current weather data.'
                  forecast:
                    type: array
                    example:
                      -
                        date: '2025-03-23'
                        temperature_max: 13.9
                        temperature_min: 7.7
                        weather_code: 80
                      -
                        date: '2025-03-24'
                        temperature_max: 15.9
                        temperature_min: 7.8
                        weather_code: 80
                      -
                        date: '2025-03-25'
                        temperature_max: 15.7
                        temperature_min: 7.6
                        weather_code: 3
                      -
                        date: '2025-03-26'
                        temperature_max: 15.9
                        temperature_min: 8.7
                        weather_code: 3
                    description: 'The weather forecast data.'
                    enum: []
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          example: '2025-03-23'
                          description: 'The date of the forecast.'
                          enum: []
                        temperature_max:
                          type: number
                          example: 13.9
                          description: 'The maximum temperature of the forecast.'
                          enum: []
                        temperature_min:
                          type: number
                          example: 7.7
                          description: 'The minimum temperature of the forecast.'
                          enum: []
                        weather_code:
                          type: integer
                          example: 80
                          description: 'The weather code of the forecast.'
                          enum: []
                    required:
                      - date
                      - temperature_max
                      - temperature_min
                      - weather_code
                required:
                  - current
                  - forecast
        400:
          description: 'Error retrieving weather data.'
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Weather data not available'
                properties:
                  error:
                    type: string
                    example: 'Weather data not available'
                    description: 'The error when retrieving weather data.'
                    enum: []
                required:
                  - error
        403:
          description: Unauthorized.
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'This action is unauthorized.'
                properties:
                  message:
                    type: string
                    example: 'This action is unauthorized.'
                    description: 'The error message.'
                    enum: []
                required:
                  - message
        404:
          description: 'Journey not found.'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'No query results for model [App\Models\Journey] 9e807de1-4b8e-4cbd-9dc5-cf2addb6a89b'
                properties:
                  message:
                    type: string
                    example: 'No query results for model [App\Models\Journey] 9e807de1-4b8e-4cbd-9dc5-cf2addb6a89b'
                    description: 'The error message.'
                    enum: []
                required:
                  - message
      tags:
        - Journey
    parameters:
      -
        in: path
        name: journey_id
        description: 'The ID of the journey.'
        example: 0023df66-7dfa-4009-afb0-57601eaa717c
        required: true
        schema:
          type: string
      -
        in: path
        name: id
        description: 'The UUID of the journey.'
        example: 9c23279f-f2a7-4666-bc00-56dfe0c47342
        required: true
        schema:
          type: uuid
...